+++ src/types/spl/playerAuthorities.ts
@@ -1,0 +1,9 @@
+export interface SplPlayerAuthorities {
+  name: string;
+  require_active_auth?: boolean;
+  authorities?: {
+    rental?: string[];
+    delegation?: string[];
+    purchase?: string[];
+  };
+}